+++ quanmiaolightapp-20240801/alibabacloud_quanmiaolightapp20240801/models/_run_video_analysis_request.py
@@ -33,6 +33,7 @@ class RunVideoAnalysisRequest(DaraModel):
         video_roles: List[main_models.RunVideoAnalysisRequestVideoRoles] = None,
         video_shot_face_identity_count: int = None,
         video_url: str = None,
+        video_urls: List[str] = None,
     ):
         self.add_document_param = add_document_param
         self.auto_role_recognition_video_url = auto_role_recognition_video_url
@@ -57,6 +58,7 @@ class RunVideoAnalysisRequest(DaraModel):
         self.video_roles = video_roles
         self.video_shot_face_identity_count = video_shot_face_identity_count
         self.video_url = video_url
+        self.video_urls = video_urls
 
     def validate(self):
         if self.add_document_param:
@@ -152,6 +154,9 @@ class RunVideoAnalysisRequest(DaraModel):
         if self.video_url is not None:
             result['videoUrl'] = self.video_url
 
+        if self.video_urls is not None:
+            result['videoUrls'] = self.video_urls
+
         return result
 
     def from_map(self, m: dict = None):
@@ -233,6 +238,9 @@ class RunVideoAnalysisRequest(DaraModel):
 
         if m.get('videoUrl') is not None:
             self.video_url = m.get('videoUrl')
+
+        if m.get('videoUrls') is not None:
+            self.video_urls = m.get('videoUrls')
 
         return self