+++ quanmiaolightapp-20240801/alibabacloud_quanmiaolightapp20240801/models/_run_video_analysis_shrink_request.py
@@ -30,6 +30,7 @@ class RunVideoAnalysisShrinkRequest(DaraModel):
         video_roles_shrink: str = None,
         video_shot_face_identity_count: int = None,
         video_url: str = None,
+        video_urls_shrink: str = None,
     ):
         self.add_document_param_shrink = add_document_param_shrink
         self.auto_role_recognition_video_url = auto_role_recognition_video_url
@@ -54,6 +55,7 @@ class RunVideoAnalysisShrinkRequest(DaraModel):
         self.video_roles_shrink = video_roles_shrink
         self.video_shot_face_identity_count = video_shot_face_identity_count
         self.video_url = video_url
+        self.video_urls_shrink = video_urls_shrink
 
     def validate(self):
         pass
@@ -132,6 +134,9 @@ class RunVideoAnalysisShrinkRequest(DaraModel):
         if self.video_url is not None:
             result['videoUrl'] = self.video_url
 
+        if self.video_urls_shrink is not None:
+            result['videoUrls'] = self.video_urls_shrink
+
         return result
 
     def from_map(self, m: dict = None):
@@ -204,6 +209,9 @@ class RunVideoAnalysisShrinkRequest(DaraModel):
 
         if m.get('videoUrl') is not None:
             self.video_url = m.get('videoUrl')
+
+        if m.get('videoUrls') is not None:
+            self.video_urls_shrink = m.get('videoUrls')
 
         return self